requirements.txt
Django==3.2.5
djangorestframework-simplejwt==4.7.2
djangorestframework==3.12.4
psycopg2-binary
django-debug-toolbar
django-extensions
Werkzeug
coverage
black
flake8